update(): Don't try to preserve row ordering if number of rows changed
This can happen after in-place filtering. If rows were removed, the
rowcol column of the modality DF does not match the information stored
in obsmap/varmap anymore and the join fails. The rowcol column is
necessary for dealing with duplicate names, since in this case the index
alone is ambiguous.

This fix is only unambiguous if there are no duplicate names present,
the behavior in the presence of duplicates is undefined.